+++ plugins/sociallogin/linkedin/linkedin.php
@@ -17,9 +17,12 @@ use Akeeba\SocialLogin\Library\Helper\Joomla;
 use Akeeba\SocialLogin\Library\Helper\Login;
 use Akeeba\SocialLogin\LinkedIn\OAuth as LinkedInOAuth;
 use Akeeba\SocialLogin\LinkedIn\UserQuery;
+use Joomla\CMS\Authentication\Authentication;
+use Joomla\CMS\Log\Log;
 use Joomla\CMS\Plugin\CMSPlugin;
 use Joomla\CMS\User\User;
 use Joomla\Registry\Registry;
+use Joomla\Utilities\ArrayHelper;
 
 if (!class_exists('Akeeba\\SocialLogin\\Library\\Helper\\Login', true))
 {
@@ -128,6 +131,9 @@ class plgSocialloginLinkedin extends CMSPlugin
 
 		// Set the integration name from the plugin name (without the plg_sociallogin_ part, of course)
 		$this->integrationName = $this->_name;
+
+		// Register a debug log file writer
+		Joomla::addLogger($this->_name);
 
 		// Load the plugin options into properties
 		$this->appId               = $this->params->get('appid', '');
@@ -378,6 +384,8 @@ class plgSocialloginLinkedin extends CMSPlugin
 	 */
 	public function onAjaxLinkedin()
 	{
+		Joomla::log($this->integrationName, 'Begin handing of authentication callback');
+
 		// This is the return URL used by the Link button
 		$returnURL  = Joomla::getSessionVar('returnUrl', JUri::base(), 'plg_system_sociallogin');
 		// And this is the login success URL used by the Login button
@@ -408,10 +416,14 @@ class plgSocialloginLinkedin extends CMSPlugin
 		{
 			try
 			{
+				Joomla::log($this->integrationName, 'Validate received token with LinkedIn', Log::INFO);
+
 				$token = $oauthConnector->authenticate();
 
 				if ($token === false)
 				{
+					Joomla::log($this->integrationName, 'Received token from LinkedIn is invalid or the user has declined application authorization', Log::ERROR);
+
 					$errorMessage = Joomla::_('PLG_SOCIALLOGIN_LINKEDIN_ERROR_NOT_LOGGED_IN_FB');
 
 					if (defined('JDEBUG') && JDEBUG)
@@ -427,6 +439,9 @@ class plgSocialloginLinkedin extends CMSPlugin
 					throw new LoginError($errorMessage);
 				}
 
+				// Get information about the user from GitHub.
+				Joomla::log($this->integrationName, 'Retrieving LinkedIn profile information', Log::INFO);
+
 				// Get information about the user from LinkedIn.
 				$tokenArray   = $oauthConnector->getToken();
 
@@ -439,10 +454,14 @@ class plgSocialloginLinkedin extends CMSPlugin
 			}
 			catch (Exception $e)
 			{
+				Joomla::log($this->integrationName, "Returning login error '{$e->getMessage()}'", Log::ERROR);
+
 				throw new LoginError($e->getMessage());
 			}
 
 			// The data used to login or create a user
+			Joomla::log($this->integrationName, sprintf("Retrieved information: %s", ArrayHelper::toString((array)$liUserFields)));
+
 			$userData = new UserData();
 			$userData->name = $liUserFields->firstName . ' ' . $liUserFields->lastName;
 			$userData->id = $liUserFields->id;
@@ -466,14 +485,18 @@ class plgSocialloginLinkedin extends CMSPlugin
 				'pictureUrl' => $liUserFields->pictureUrl,
 			);
 
+			Joomla::log($this->integrationName, sprintf("Calling Social Login login handler with the following information: %s", ArrayHelper::toString($userProfileData)));
+
 			Login::handleSocialLogin($this->integrationName, $pluginConfiguration, $userData, $userProfileData);
 		}
 		catch (LoginError  $e)
 		{
 			// Log failed login
 			$response                = Login::getAuthenticationResponseObject();
-			$response->status        = JAuthentication::STATUS_UNKNOWN;
+			$response->status        = Authentication::STATUS_UNKNOWN;
 			$response->error_message = $e->getMessage();
+
+			Joomla::log($this->integrationName, sprintf("Received login failure. Message: %s", $e->getMessage()), Log::ERROR);
 
 			// This also enqueues the login failure message for display after redirection. Look for JLog in that method.
 			Login::processLoginFailure($response);
@@ -484,6 +507,8 @@ class plgSocialloginLinkedin extends CMSPlugin
 		}
 		catch (GenericMessage $e)
 		{
+			Joomla::log($this->integrationName, sprintf("Report non-login failure message to user: %s", $e->getMessage()), Log::NOTICE);
+
 			// Do NOT go through processLoginFailure. This is NOT a login failure.
 			$app->enqueueMessage($e->getMessage(), 'info');
 			$app->redirect($failureUrl);
@@ -491,6 +516,7 @@ class plgSocialloginLinkedin extends CMSPlugin
 			return;
 		}
 
+		Joomla::log($this->integrationName, sprintf("Successful login. Redirecting to %s", $loginUrl), Log::INFO);
 		$app->redirect($loginUrl);
 	}