+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **New features**
 
 * Login with Google using Google OpenID instead of the now defunct Google+ API
+* Debug logs for all Social Login plugins
 
 # 2.0.3